+++ packages/charts/react-charts/stories/src/AreaChart/AreaChartNegative..stories.tsx
@@ -1,6 +1,20 @@
 import * as React from 'react';
 import { AreaChart } from '@fluentui/react-charts';
-import { Switch, Field, Radio, RadioGroup, RadioGroupOnChangeData } from '@fluentui/react-components';
+import {
+  Switch,
+  Field,
+  Radio,
+  RadioGroup,
+  RadioGroupOnChangeData,
+  makeStyles,
+  tokens,
+} from '@fluentui/react-components';
+
+const useStyles = makeStyles({
+  svgTooltip: {
+    fill: tokens.colorNeutralBackground2,
+  },
+});
 
 export const AreaChartNegative = () => {
   const [width, setWidth] = React.useState<number>(700);
@@ -8,6 +22,7 @@ export const AreaChartNegative = () => {
   const [isCalloutSelected, setIsCalloutSelected] = React.useState<boolean>(false);
   const [showAxisTitles, setShowAxisTitles] = React.useState<boolean>(true);
 
+  const classes = useStyles();
   React.useEffect(() => {
     const style = document.createElement('style');
     const focusStylingCSS = `
@@ -199,6 +214,7 @@ export const AreaChartNegative = () => {
             enablePerfOptimization={true}
             yAxisTitle={showAxisTitles ? 'Variation of stock market prices' : undefined}
             xAxisTitle={showAxisTitles ? 'Number of days' : undefined}
+            styles={{ svgTooltip: classes.svgTooltip }}
           />
         </div>
       )}
@@ -210,6 +226,7 @@ export const AreaChartNegative = () => {
             width={width}
             data={chartData}
             enablePerfOptimization={true}
+            styles={{ svgTooltip: classes.svgTooltip }}
           />
         </div>
       )}