+++ packages/charts/react-charts/stories/src/VerticalStackedBarChart/VerticalStackedBarChartSecondaryYAxis.stories.tsx
@@ -7,9 +7,16 @@ import {
   DataVizPalette,
   getColorFromToken,
 } from '@fluentui/react-charts';
-import { useId } from '@fluentui/react-components';
+import { useId, makeStyles, tokens } from '@fluentui/react-components';
+
+const useStyles = makeStyles({
+  svgTooltip: {
+    fill: tokens.colorNeutralBackground2,
+  },
+});
 
 export const VerticalStackedBarSecondaryYAxis = () => {
+  const classes = useStyles();
   const [width, setWidth] = React.useState<number>(700);
   const [height, setHeight] = React.useState<number>(300);
   const _widthSliderId = useId('width-slider-');
@@ -221,6 +228,7 @@ export const VerticalStackedBarSecondaryYAxis = () => {
           yAxisTitle="Variation of number of sales"
           xAxisTitle="Number of days"
           secondaryYScaleOptions={{}}
+          styles={{ svgTooltip: classes.svgTooltip }}
         />
       </div>
     </div>