+++ packages/fluentui/react-northstar-styles-renderer/src/index.ts
@@ -1,2 +1,0 @@
-export { noopRenderer } from './noopRenderer';
-export * from './types';