+++ packages/fluentui/react-northstar/bundle-size/Chat.fixture.js
@@ -1,7 +1,0 @@
-import { Chat } from '@fluentui/react-northstar';
-
-console.log(Chat);
-
-export default {
-  name: 'Chat',
-};