+++ packages/fluentui/react-northstar/bundle-size/Grid.fixture.js
@@ -1,7 +1,0 @@
-import { Grid } from '@fluentui/react-northstar';
-
-console.log(Grid);
-
-export default {
-  name: 'Grid',
-};