+++ packages/fluentui/react-northstar/src/themes/teams/components/Table/tableRowVariables.ts
@@ -1,1 +1,0 @@
-export { tableVariables as tableRowVariables } from './tableVariables';