+++ packages/fluentui/react-northstar/test/specs/components/List/ListItemHeaderMedia-test.tsx
@@ -1,6 +1,0 @@
-import { ListItemHeaderMedia } from 'src/components/List/ListItemHeaderMedia';
-import { isConformant } from 'test/specs/commonTests';
-
-describe('ListItemHeaderMedia', () => {
-  isConformant(ListItemHeaderMedia, { testPath: __filename, constructorName: 'ListItemHeaderMedia' });
-});