+++ packages/react-components/react-menu/library/src/components/Menu/useMenu.tsx
@@ -44,7 +44,7 @@ const submenuFallbackPositions: PositioningShorthandValue[] = [
  *
  * @param props - props from this instance of Menu
  */
-export const useMenu_unstable = (props: MenuProps & { safeZone?: boolean }): MenuState => {
+export const useMenu_unstable = (props: MenuProps & { safeZone?: boolean | { timeout?: number } }): MenuState => {
   const isSubmenu = useIsSubmenu();
   const {
     hoverDelay = 500,
@@ -105,7 +105,7 @@ export const useMenu_unstable = (props: MenuProps & { safeZone?: boolean }): Men
 
   const safeZoneHandle = useSafeZoneArea({
     disabled: !enableSafeZone,
-    timeout: 1500,
+    timeout: typeof safeZone === 'object' ? safeZone.timeout : 300,
 
     onSafeZoneEnter: e => {
       setOpen(e, { open: true, keyboard: false, type: 'menuSafeZoneMouseEnter', event: e });
@@ -260,16 +260,20 @@ const useMenuOpenState = (
 
   const [setOpenTimeout, clearOpenTimeout] = useTimeout();
 
-  const setOpen = useEventCallback((e: MenuOpenEvent, data: MenuOpenChangeData) => {
+  const setOpen = useEventCallback((e: MenuOpenEvent, data: MenuOpenChangeData & { ignoreHoverDelay?: boolean }) => {
     clearOpenTimeout();
     if (!(e instanceof Event) && e.persist) {
       // < React 17 still uses pooled synthetic events
       e.persist();
     }
 
-    if (e.type === 'mouseleave' || e.type === 'mouseenter' || e.type === 'mousemove' || e.type === MENU_ENTER_EVENT) {
+    const shouldUseDelay =
+      !data.ignoreHoverDelay &&
+      (e.type === 'mouseleave' || e.type === 'mouseover' || e.type === 'mousemove' || e.type === MENU_ENTER_EVENT);
+
+    if (shouldUseDelay) {
       if (state.triggerRef.current?.contains(e.target as HTMLElement)) {
-        enteringTriggerRef.current = e.type === 'mouseenter' || e.type === 'mousemove';
+        enteringTriggerRef.current = e.type === 'mouseover' || e.type === 'mousemove';
       }
 
       setOpenTimeout(() => trySetOpen(e, data), state.hoverDelay);