+++ packages/react-components/react-combobox/library/CHANGELOG.md
@@ -1,8 +1,18 @@
 # Change Log - @fluentui/react-combobox
 
-This log was last generated on Fri, 04 Jul 2025 10:02:42 GMT and should not be manually modified.
+This log was last generated on Fri, 11 Jul 2025 15:59:24 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.15.7](https://github.com/microsoft/fluentui/tree/@fluentui/react-combobox_v9.15.7)
+
+Fri, 11 Jul 2025 15:59:24 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-combobox_v9.15.6..@fluentui/react-combobox_v9.15.7)
+
+### Patches
+
+- Bump @fluentui/react-field to v9.3.7 ([PR #34807](https://github.com/microsoft/fluentui/pull/34807) by beachball)
+- Bump @fluentui/react-positioning to v9.19.0 ([PR #34807](https://github.com/microsoft/fluentui/pull/34807) by beachball)
 
 ## [9.15.6](https://github.com/microsoft/fluentui/tree/@fluentui/react-combobox_v9.15.6)